baseline check - root cause is dask versioning
The prefect 3 API really needs a newer version of dask in order to run
effectively. The issues have work arounds, but they are ugle and
ultimately unmaintainable.

IN th short term the best course is likely to do nothing.

In the mid- to long-term the options are to either drop usage of
crystalball, which relies on dask-ms which is where the pin to an older
dask version comes from, try out some of their early PRs.